Move projection to a separate subclause
