+++ packages/graphql/src/translate/cypher-builder/CypherBuilder.ts
@@ -22,10 +22,10 @@ export { Match } from "./clauses/Match";
 export { Create } from "./clauses/Create";
 export { Merge } from "./clauses/Merge";
 export { Call } from "./clauses/Call";
-export { Return } from "./clauses/projection/Return";
+export { Return } from "./clauses/Return";
 export { RawCypher } from "./clauses/RawCypher";
-export { With } from "./clauses/projection/With";
-export { Unwind } from "./clauses/projection/Unwind";
+export { With } from "./clauses/With";
+export { Unwind } from "./clauses/Unwind";
 
 // Clauses-like things
 export { Exists } from "./Exists";
